Add alternative day 11 solution using summed-area table
The usage of a pre-computed value table (here: summed-area) speeds up the compution extremly as we can reduce the complecity down to a `O(n^3)` where `n` is the size of the grid (here `300`). Before that, it was at least `O(n^5)`.

Inspired by https://todd.ginsberg.com/post/advent-of-code/2018/day11/